Tests/LibWeb/Text/input/HTML/HTMLMediaElement-load-after-decode-error.html

<!DOCTYPE html>
<script src="../include.js"></script>
<video id="video"></video>
<script>
    const video = document.getElementById("video");

    function logEvent(name) {
        let entry = `${name}: readyState=${video.readyState}`;
        if (name === "durationchange")
            entry += ` duration=${video.duration}`;
        if (name === "error")
            entry += ` code=${video.error.code}`;
        println(entry);
    }

    for (const name of [
        "loadstart", "abort", "error", "emptied",
        "loadedmetadata", "loadeddata", "canplay",
        "canplaythrough", "durationchange", "ended",
    ]) {
        video.addEventListener(name, () => logEvent(name));
    }

    asyncTest(async done => {
        video.src = "../../../Assets/corrupt-video.webm";

        await new Promise(resolve => video.addEventListener("error", resolve, { once: true }));
        video.src = "../../../Assets/test-webm.webm";

        await new Promise(resolve => video.addEventListener("canplaythrough", resolve, { once: true }));
        done();
    });
</script>
