Tests/LibWeb/Text/expected/HTML/HTMLMediaElement-load-after-decode-error.txt

loadstart: readyState=0
durationchange: readyState=1 duration=0
loadedmetadata: readyState=1
error: readyState=1 code=3
abort: readyState=0
emptied: readyState=0
loadstart: readyState=0
durationchange: readyState=2 duration=12.043
loadedmetadata: readyState=2
loadeddata: readyState=2
canplay: readyState=4
canplaythrough: readyState=4
